move meta and common schema to Context
